Removes unused image assets
Removes a large number of image assets that are no longer used by the site. This reduces the size of the repository and the deployed site.

These images include various sizes of favicons, touch icons, cover images for old blog posts and weeknotes, and character images.

The `make-post.py` script is adjusted to correctly format the year in post paths.